pyproject.toml
[tool.poetry]
name = "cli-base"
version = "0.1.0"
description = "A simple generic Web App CLI"
authors = ["Mark Hall <mark.hall@work.room3b.eu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
PyYAML = "^6.0"
Cerberus = "^1.3.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
flake8-print = "^5.0.0"
flake8-docstrings = "^1.6.0"
flake8-annotations = "^2.9.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
tox = "^3.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
